requirements.txt
```
fastapi
uvicorn
pandas
numpy
scikit-learn==1.7.2
lightgbm==4.6.0
xgboost==3.2.0
catboost==1.2.10
psycopg2-binary
joblib
faker
plotly
requests
pydantic
python-multipart